论文管理》指定校优，查看校优，指定省优

+++ src/com/newview/bysj/web/provinceExcellentProject/AssignProvinceExcellentProject.java
@@ -33,7 +33,7 @@ public class AssignProvinceExcellentProject extends BaseController {
         PageInfo pageInfo = new PageInfo();
         Page<GraduateProject> graduateProject = graduateProjectService.getPagesForProvinceExcellentCandidate(title == null ? null : title, tutorName == null ? null : tutorName, page, rows);
         pageInfo.setRows(graduateProject.getContent());
-        pageInfo.setTotal((int) graduateProject.getNumberOfElements());
+        pageInfo.setTotal(graduateProject.getNumberOfElements());
         return pageInfo;
     }
